feat: Multi-Notensystem (DE 0-15, DE 1-6, CH, AT, IB) (#3)
* refactor: Notensystem-Interface erweitern, calc system-agnostisch (DE-Default)


* refactor: Notensystem-Interface erweitern, calc system-agnostisch + 4 neue Systeme

- systems.ts: DE_1_6, CH_1_6, AT_1_5, IB_1_7 + ALLE_SYSTEME-Export
- calc.ts: clamp(wert, system) generisch, alle Schnitt-Fns system-aware (DE-Default)
- schnitt-farbe.ts: delegiert an system.farbe (richtungsabhängig)
- 79 Tests grün


* fix: error handling in API routes and server pages

- coach/route.ts: try-catch um req.json() und Anthropic API call — gibt 400 bzw. 503 zurück statt unkontrolliert zu crashen
- export/route.ts: Promise.all in try-catch; Supabase-Errors werden geprüft und geloggt statt still ignoriert
- dashboard/noten/einstellungen pages: alle Supabase-Reads extrahieren jetzt { error } und loggen via console.error — Fehler landen in Vercel-Logs statt unter den ?? []-Fallbacks zu verschwinden


* feat(db): punkte numeric(4,2) + notensystem-CHECK fuer alle 5 Systeme

- schule_note.punkte: smallint -> numeric(4,2) (CH-Kommanoten)
- nutzer_profil.notensystem_check: de_0_15 | de_1_6 | ch_1_6 | at_1_5 | ib_1_7
- TS-Typen regeneriert


* feat: Noten-Actions system-aware + setNotensystem + noteAnzahl

- AddNoteSchema.punkte: .int() entfernt (CH-Kommas erlaubt)
- addNote/updateNote: system-aware Validierung, roh speichern (kein Math.round)
- userNotensystem() Helper laedt aktives System aus nutzer_profil
- setNotensystem() + noteAnzahl() Actions neu


* feat: NotensystemProvider + useNotensystem-Hook

Provider laedt notensystem aus nutzer_profil im App-Layout und
stellt es per Context bereit. Alle Client-Komponenten koennen
useNotensystem() nutzen.


* chore: package-updates + console.log aus cron-route entfernen

Package-Updates (Patch/Minor, nicht-geschützt):
- @anthropic-ai/sdk 0.100.1 → 0.104.1
- @tailwindcss/postcss 4.3.0 → 4.3.1
- @types/node 20.19.41 → 20.19.43
- @types/react 19.2.15 → 19.2.17
- lucide-react 1.16.0 → 1.18.0

PostCSS XSS (GHSA-qx2v-qp2m-jg93, moderate): kein Fix ohne
Next.js-Major-Downgrade — kein User-Generated-CSS im Projekt,
Risiko effektiv null.


* refactor: Noten-Komponenten nutzen useNotensystem statt fester DE-Anzeige

- schnitt-karte.tsx: system.formatNote, system.max, normierte Progressbar
- notenrechner-board.tsx: gesamtSchnittGerundet(system), system.formatNote
- jahres-tabelle.tsx: schnittFarbe(system), system.formatNote, system.max
- fach-card.tsx: KategorienSplit + SchnittVerlauf normiert, system.parse in AddNote/NoteEditForm
- was-waere-wenn-panel.tsx: system-aware Inputs, fachSchnittGerundet(system), benoetigtePunkte(system)
- was-waere-wenn-seite.tsx: SchnittBalken normiert, gesamtSchnittGerundet(system), dynamische Quick-Picks


* feat: Notensystem-Wahl in Einstellungen (Schule-Sektion)

- NotensystemWahl-Dropdown mit allen 5 Systemen (DE/CH/AT/IB)
- Sofort-Save via setNotensystem-Action, kein Page-Reload nötig
- Eingebettet in /einstellungen unter Bundesland (Schule-Block)
- notensystem-Feld in Profil-Select ergänzt


* fix: router.refresh() nach Notensystem-Wechsel damit Layout neu lädt


---------

+++ lib/validation.ts
@@ -20,7 +20,7 @@ export const AddFachSchema = z.object({
 
 export const AddNoteSchema = z.object({
   fachId: FachIdSchema,
-  punkte: z.number().int().min(0).max(15),
+  punkte: z.number().min(0).max(15), // .int() entfällt (CH-Kommas); Per-System-Check in der Action
   kategorie: KategorieSchema,
   bezeichnung: z.string().max(200).optional(),
   gewicht: z.number().positive().max(10).optional(),
